+++ src/ParserInterface.test.js
@@ -44,18 +44,3 @@ describe("Given new ParserInterface has been created", () => {
     expect(parser.finalResults).toEqual(null);
   });
 });
-
-describe("Given new ParserInterface was created", () => {
-  const parser = new ParserInterface();
-
-  describe("when incrementId is passed a parameter returns a string", () => {
-    const result = parser.incrementId(9);
-    expect(typeof result).toEqual("string");
-  });
-
-  describe("when incrementId is passed a parameter it is turned into a string", () => {
-    const result = parser.incrementId(9.5);
-    expect(typeof result).toEqual("string");
-    expect(result).toEqual("9.6");
-  });
-});
